+++ examples/tutorial-examples/AreaSurfaceEstimation-template.cpp
@@ -1,5 +1,6 @@
 /// @file tutorial-examples/AreaSurfaceEstimation-template.cpp
 /// Shape
+
 #include "DGtal/shapes/parametric/Ball3D.h"
 
 //! [areaSurface-basicIncludes]